refactor(*): move common DB tables to models package
