+++ src/ramstk/dao/DAO.py
@@ -16,7 +16,12 @@ from sqlalchemy.orm import scoped_session, sessionmaker
 
 # RAMSTK Package Imports
 from ramstk.exceptions import DataAccessError
-from ramstk.models.commondb import RAMSTKSiteInfo
+from ramstk.models.commondb import (
+    RAMSTKRPN, RAMSTKCategory, RAMSTKCondition, RAMSTKFailureMode, RAMSTKGroup,
+    RAMSTKHazards, RAMSTKLoadHistory, RAMSTKManufacturer, RAMSTKMeasurement,
+    RAMSTKMethod, RAMSTKModel, RAMSTKSiteInfo, RAMSTKStakeholders,
+    RAMSTKStatus, RAMSTKSubCategory, RAMSTKType, RAMSTKUser
+)
 from ramstk.models.programdb import (
     RAMSTKNSWC, RAMSTKAction, RAMSTKAllocation, RAMSTKCause, RAMSTKControl,
     RAMSTKDesignElectric, RAMSTKDesignMechanic, RAMSTKEnvironment,
@@ -29,12 +34,6 @@ from ramstk.models.programdb import (
 )
 
 # RAMSTK Local Imports
-from .commondb import (
-    RAMSTKRPN, RAMSTKCategory, RAMSTKCondition, RAMSTKFailureMode,
-    RAMSTKGroup, RAMSTKHazards, RAMSTKLoadHistory, RAMSTKManufacturer,
-    RAMSTKMeasurement, RAMSTKMethod, RAMSTKModel, RAMSTKStakeholders,
-    RAMSTKStatus, RAMSTKSubCategory, RAMSTKType, RAMSTKUser
-)
 from .RAMSTKCommonDB import (
     RAMSTK_CATEGORIES, RAMSTK_CONDITIONS, RAMSTK_FAILURE_MODES,
     RAMSTK_GROUPS, RAMSTK_HAZARDS, RAMSTK_HISTORIES, RAMSTK_MANUFACTURERS,
