+++ tests/models/commondb/test_ramstkcondition.py
@@ -1,44 +1,43 @@
 # -*- coding: utf-8 -*-
 #
-#       tests.dao.commondb.test_ramstkfailuremode.py is part of The RAMSTK Project
+#       tests.dao.commondb.test_ramstkcondition.py is part of The RAMSTK Project
 #
 # All rights reserved.
-"""Test class for the RAMSTKFailureMode module algorithms and models. """
+"""Test class for testing the RAMSTKCondition module algorithms and models."""
 
 # Third Party Imports
 import pytest
 
 # RAMSTK Package Imports
-from ramstk.dao.commondb.RAMSTKFailureMode import RAMSTKFailureMode
+from ramstk.models.commondb import RAMSTKCondition
+
+__author__ = 'Doyle Rowland'
+__email__ = 'doyle.rowland@reliaqual.com'
+__organization__ = 'ReliaQual Associates, LLC'
+__copyright__ = 'Copyright 2017 Doyle "weibullguy" Rowland'
 
 ATTRIBUTES = {
-    'mode_id': 3,
-    'description': 'Parameter Change',
-    'subcategory_id': 24,
-    'source': 'FMD-97',
-    'category_id': 3,
-    'mode_ratio': 0.2,
+    'condition_type': 'operating',
+    'condition_id': 1,
+    'description': 'Cavitation',
 }
 
 
 @pytest.mark.integration
-def test_ramstkfailuremode_create(test_common_dao):
-    """ __init__() should create an RAMSTKFailureMode model. """
+def test_ramstkcondition_create(test_common_dao):
+    """ __init__() should create an RAMSTKCondition model. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False,
     )
-    DUT = _session.query(RAMSTKFailureMode).first()
+    DUT = _session.query(RAMSTKCondition).first()
 
-    assert isinstance(DUT, RAMSTKFailureMode)
+    assert isinstance(DUT, RAMSTKCondition)
 
     # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_failure_mode'
-    assert DUT.category_id == 3
-    assert DUT.subcategory_id == 24
-    assert DUT.mode_id == 3
-    assert DUT.description == 'Parameter Change'
-    assert DUT.mode_ratio == 0.2
-    assert DUT.source == 'FMD-97'
+    assert DUT.__tablename__ == 'ramstk_condition'
+    assert DUT.condition_id == 1
+    assert DUT.description == 'Cavitation'
+    assert DUT.cond_type == 'operating'
 
 
 @pytest.mark.integration
@@ -47,7 +46,7 @@ def test_get_attributes(test_common_dao):
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False,
     )
-    DUT = _session.query(RAMSTKFailureMode).first()
+    DUT = _session.query(RAMSTKCondition).first()
 
     assert DUT.get_attributes() == ATTRIBUTES
 
@@ -58,14 +57,14 @@ def test_set_attributes(test_common_dao):
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False,
     )
-    DUT = _session.query(RAMSTKFailureMode).first()
+    DUT = _session.query(RAMSTKCondition).first()
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 0
     assert _msg == (
-        "RAMSTK SUCCESS: Updating RAMSTKFailureMode {0:d} "
-        "attributes.".format(DUT.mode_id)
+        "RAMSTK SUCCESS: Updating RAMSTKCondition {0:d} "
+        "attributes.".format(DUT.condition_id)
     )
 
 
@@ -75,16 +74,17 @@ def test_set_attributes_missing_key(test_common_dao):
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False,
     )
-    DUT = _session.query(RAMSTKFailureMode).first()
+    DUT = _session.query(RAMSTKCondition).first()
 
-    ATTRIBUTES.pop('mode_ratio')
+    ATTRIBUTES.pop('condition_type')
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 40
     assert _msg == (
-        "RAMSTK ERROR: Missing attribute 'mode_ratio' in attribute "
-        "dictionary passed to RAMSTKFailureMode.set_attributes()."
+        "RAMSTK ERROR: Missing attribute 'condition_type' in "
+        "attribute dictionary passed to "
+        "RAMSTKCondition.set_attributes()."
     )
 
-    ATTRIBUTES['mode_ratio'] = 0.2
+    ATTRIBUTES['condition_type'] = 'operating'
