+++ tests/models/commondb/test_ramstkgroup.py
@@ -1,40 +1,41 @@
+#!/usr/bin/env python -O
 # -*- coding: utf-8 -*-
 #
-#       tests.dao.commondb.test_ramstkmodel.py is part of The RAMSTK Project
+#       tests.unit._dao.TestRAMSTKUser.py is part of The RAMSTK Project
+
 #
 # All rights reserved.
-"""Test class for testing the RAMSTKModel module algorithms and models."""
+"""Test class for testing the RAMSTKUser module algorithms and models."""
 
+# Third Party Imports
 import pytest
 
-from ramstk.dao.commondb.RAMSTKModel import RAMSTKModel
+# RAMSTK Package Imports
+from ramstk.models.commondb import RAMSTKGroup
 
 __author__ = 'Doyle Rowland'
 __email__ = 'doyle.rowland@reliaqual.com'
 __organization__ = 'ReliaQual Associates, LLC'
 __copyright__ = 'Copyright 2017 Doyle "weibullguy" Rowland'
 
-ATTRIBUTES = {
-    'model_type': 'damage',
-    'model_id': 1,
-    'description': 'Adhesion Wear Model for Bearings'
-}
+
+ATTRIBUTES = {'group_id': 1, 'description':'Engineering, Design', 'group_type':'workgroup'}
 
 
 @pytest.mark.integration
-def test_ramstkmodel_create(test_common_dao):
-    """ __init__() should create an RAMSTKModel model. """
+def test_ramstkworkgroup_create(test_common_dao):
+    """ __init__() should create an RAMSTKGroup model. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKModel).first()
+    DUT = _session.query(RAMSTKGroup).first()
 
-    assert isinstance(DUT, RAMSTKModel)
+    assert isinstance(DUT, RAMSTKGroup)
 
     # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_model'
-    assert DUT.model_id == 1
-    assert DUT.description == 'Adhesion Wear Model for Bearings'
-    assert DUT.model_type == 'damage'
+    assert DUT.__tablename__ == 'ramstk_group'
+    assert DUT.group_id == 1
+    assert DUT.description == 'Engineering, Design'
+    assert DUT.group_type == 'workgroup'
 
 
 @pytest.mark.integration
@@ -42,7 +43,7 @@ def test_get_attributes(test_common_dao):
     """ get_attributes() should return a tuple of attributes values on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKModel).first()
+    DUT = _session.query(RAMSTKGroup).first()
 
     assert DUT.get_attributes() == ATTRIBUTES
 
@@ -52,13 +53,13 @@ def test_set_attributes(test_common_dao):
     """ set_attributes() should return a zero error code on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKModel).first()
+    DUT = _session.query(RAMSTKGroup).first()
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 0
-    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKModel {0:d} "
-                    "attributes.".format(DUT.model_id))
+    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKGroup {0:d} "
+                    "attributes.".format(DUT.group_id))
 
 
 @pytest.mark.integration
@@ -66,14 +67,14 @@ def test_set_attributes_missing_key(test_common_dao):
     """ set_attributes() should return a 40 error code when passed too few attributes. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKModel).first()
+    DUT = _session.query(RAMSTKGroup).first()
 
-    ATTRIBUTES.pop('model_type')
+    ATTRIBUTES.pop('group_type')
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 40
-    assert _msg == ("RAMSTK ERROR: Missing attribute 'model_type' in attribute "
-                    "dictionary passed to RAMSTKModel.set_attributes().")
+    assert _msg == ("RAMSTK ERROR: Missing attribute 'group_type' in attribute "
+                    "dictionary passed to RAMSTKGroup.set_attributes().")
 
-    ATTRIBUTES['model_type'] = 'damage'
+    ATTRIBUTES['group_type'] = 'workgroup'
