+++ tests/models/commondb/test_ramstkmodel.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 #
-#       tests.dao.commondb.test_ramstkstatus.py is part of The RAMSTK Project
+#       tests.dao.commondb.test_ramstkmodel.py is part of The RAMSTK Project
 #
 # All rights reserved.
-"""Test class for testing the RAMSTKStatus module algorithms and models."""
+"""Test class for testing the RAMSTKModel module algorithms and models."""
 
+# Third Party Imports
 import pytest
 
-from ramstk.dao.commondb.RAMSTKStatus import RAMSTKStatus
+# RAMSTK Package Imports
+from ramstk.models.commondb import RAMSTKModel
 
 __author__ = 'Doyle Rowland'
 __email__ = 'doyle.rowland@reliaqual.com'
@@ -15,28 +17,26 @@ __organization__ = 'ReliaQual Associates, LLC'
 __copyright__ = 'Copyright 2017 Doyle "weibullguy" Rowland'
 
 ATTRIBUTES = {
-    'description': 'Incident has been initiated.',
-    'name': 'Initiated',
-    'status_type': 'incident',
-    'status_id': 1
+    'model_type': 'damage',
+    'model_id': 1,
+    'description': 'Adhesion Wear Model for Bearings'
 }
 
 
 @pytest.mark.integration
-def test_ramstkstatus_create(test_common_dao):
-    """ __init__() should create an RAMSTKStatus model. """
+def test_ramstkmodel_create(test_common_dao):
+    """ __init__() should create an RAMSTKModel model. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKStatus).first()
+    DUT = _session.query(RAMSTKModel).first()
 
-    assert isinstance(DUT, RAMSTKStatus)
+    assert isinstance(DUT, RAMSTKModel)
 
     # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_status'
-    assert DUT.status_id == 1
-    assert DUT.name == 'Initiated'
-    assert DUT.description == 'Incident has been initiated.'
-    assert DUT.status_type == 'incident'
+    assert DUT.__tablename__ == 'ramstk_model'
+    assert DUT.model_id == 1
+    assert DUT.description == 'Adhesion Wear Model for Bearings'
+    assert DUT.model_type == 'damage'
 
 
 @pytest.mark.integration
@@ -44,7 +44,7 @@ def test_get_attributes(test_common_dao):
     """ get_attributes() should return a tuple of attributes values on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKStatus).first()
+    DUT = _session.query(RAMSTKModel).first()
 
     assert DUT.get_attributes() == ATTRIBUTES
 
@@ -54,13 +54,13 @@ def test_set_attributes(test_common_dao):
     """ set_attributes() should return a zero error code on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKStatus).first()
+    DUT = _session.query(RAMSTKModel).first()
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 0
-    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKStatus {0:d} "
-                    "attributes.".format(DUT.status_id))
+    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKModel {0:d} "
+                    "attributes.".format(DUT.model_id))
 
 
 @pytest.mark.integration
@@ -68,14 +68,14 @@ def test_set_attributes_missing_key(test_common_dao):
     """ set_attributes() should return a 40 error code when passed too few attributes. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKStatus).first()
+    DUT = _session.query(RAMSTKModel).first()
 
-    ATTRIBUTES.pop('name')
+    ATTRIBUTES.pop('model_type')
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 40
-    assert _msg == ("RAMSTK ERROR: Missing attribute 'name' in attribute "
-                    "dictionary passed to RAMSTKStatus.set_attributes().")
+    assert _msg == ("RAMSTK ERROR: Missing attribute 'model_type' in attribute "
+                    "dictionary passed to RAMSTKModel.set_attributes().")
 
-    ATTRIBUTES['name'] = ''
+    ATTRIBUTES['model_type'] = 'damage'
