+++ tests/models/commondb/test_ramstkstakeholders.py
@@ -1,50 +1,45 @@
 # -*- coding: utf-8 -*-
 #
-#       tests.dao.commondb.test_ramstkmethod.py is part of The RAMSTK Project
+#       tests.dao.commondb.test_ramstkstakeholders.py is part of The RAMSTK Project
 #
 # All rights reserved.
-"""Test class for testing the RAMSTKMethod module algorithms and models."""
+"""Test class for testing the RAMSTKStakeholders module algorithms and models."""
 
+# Third Party Imports
 import pytest
 
-from ramstk.dao.commondb.RAMSTKMethod import RAMSTKMethod
+# RAMSTK Package Imports
+from ramstk.models.commondb import RAMSTKStakeholders
 
 __author__ = 'Doyle Rowland'
 __email__ = 'doyle.rowland@reliaqual.com'
 __organization__ = 'ReliaQual Associates, LLC'
 __copyright__ = 'Copyright 2017 Doyle "weibullguy" Rowland'
 
-ATTRIBUTES = {
-    'method_type': 'detection',
-    'method_id': 1,
-    'name': 'Code Reviews',
-    'description': ''
-}
+ATTRIBUTES = {'stakeholders_id': 1, 'stakeholder': 'Customer'}
 
 
 @pytest.mark.integration
-def test_ramstkmethod_create(test_common_dao):
-    """ __init__() should create an RAMSTKMethod model. """
+def test_ramstkstakeholders_create(test_common_dao):
+    """ __init__() should create an RAMSTKStakeholders model. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKMethod).first()
+    DUT = _session.query(RAMSTKStakeholders).first()
 
-    assert isinstance(DUT, RAMSTKMethod)
+    assert isinstance(DUT, RAMSTKStakeholders)
 
     # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_method'
-    assert DUT.method_id == 1
-    assert DUT.description == ''
-    assert DUT.name == 'Code Reviews'
-    assert DUT.method_type == 'detection'
+    assert DUT.__tablename__ == 'ramstk_stakeholders'
+    assert DUT.stakeholders_id == 1
+    assert DUT.stakeholder == 'Customer'
 
 
 @pytest.mark.integration
 def test_get_attributes(test_common_dao):
-    """ get_attributes() should return a dict of attributes values on success. """
+    """ get_attributes() should return a tuple of attributes values on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKMethod).first()
+    DUT = _session.query(RAMSTKStakeholders).first()
 
     assert DUT.get_attributes() == ATTRIBUTES
 
@@ -54,13 +49,13 @@ def test_set_attributes(test_common_dao):
     """ set_attributes() should return a zero error code on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKMethod).first()
+    DUT = _session.query(RAMSTKStakeholders).first()
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 0
-    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKMethod {0:d} "
-                    "attributes.".format(DUT.method_id))
+    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKStakeholders {0:d} "
+                    "attributes.".format(DUT.stakeholders_id))
 
 
 @pytest.mark.integration
@@ -68,14 +63,14 @@ def test_set_attributes_missing_key(test_common_dao):
     """ set_attributes() should return a 40 error code when passed too few attributes. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKMethod).first()
+    DUT = _session.query(RAMSTKStakeholders).first()
 
-    ATTRIBUTES.pop('method_type')
+    ATTRIBUTES.pop('stakeholder')
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 40
-    assert _msg == ("RAMSTK ERROR: Missing attribute 'method_type' in attribute "
-                    "dictionary passed to RAMSTKMethod.set_attributes().")
+    assert _msg == ("RAMSTK ERROR: Missing attribute 'stakeholder' in attribute "
+                    "dictionary passed to RAMSTKStakeholders.set_attributes().")
 
-    ATTRIBUTES['method_type'] = 'test'
+    ATTRIBUTES['stakeholder'] = 'Customer'
