+++ tests/models/commondb/test_ramstkstatus.py
@@ -1,39 +1,44 @@
-#!/usr/bin/env python -O
 # -*- coding: utf-8 -*-
 #
-#       tests.unit._dao.TestRAMSTKUser.py is part of The RAMSTK Project
-
+#       tests.dao.commondb.test_ramstkstatus.py is part of The RAMSTK Project
 #
 # All rights reserved.
-"""Test class for testing the RAMSTKUser module algorithms and models."""
+"""Test class for testing the RAMSTKStatus module algorithms and models."""
 
+# Third Party Imports
 import pytest
 
-from ramstk.dao.commondb.RAMSTKGroup import RAMSTKGroup
+# RAMSTK Package Imports
+from ramstk.models.commondb import RAMSTKStatus
 
 __author__ = 'Doyle Rowland'
 __email__ = 'doyle.rowland@reliaqual.com'
 __organization__ = 'ReliaQual Associates, LLC'
 __copyright__ = 'Copyright 2017 Doyle "weibullguy" Rowland'
 
-
-ATTRIBUTES = {'group_id': 1, 'description':'Engineering, Design', 'group_type':'workgroup'}
+ATTRIBUTES = {
+    'description': 'Incident has been initiated.',
+    'name': 'Initiated',
+    'status_type': 'incident',
+    'status_id': 1
+}
 
 
 @pytest.mark.integration
-def test_ramstkworkgroup_create(test_common_dao):
-    """ __init__() should create an RAMSTKGroup model. """
+def test_ramstkstatus_create(test_common_dao):
+    """ __init__() should create an RAMSTKStatus model. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKGroup).first()
+    DUT = _session.query(RAMSTKStatus).first()
 
-    assert isinstance(DUT, RAMSTKGroup)
+    assert isinstance(DUT, RAMSTKStatus)
 
     # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_group'
-    assert DUT.group_id == 1
-    assert DUT.description == 'Engineering, Design'
-    assert DUT.group_type == 'workgroup'
+    assert DUT.__tablename__ == 'ramstk_status'
+    assert DUT.status_id == 1
+    assert DUT.name == 'Initiated'
+    assert DUT.description == 'Incident has been initiated.'
+    assert DUT.status_type == 'incident'
 
 
 @pytest.mark.integration
@@ -41,7 +46,7 @@ def test_get_attributes(test_common_dao):
     """ get_attributes() should return a tuple of attributes values on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKGroup).first()
+    DUT = _session.query(RAMSTKStatus).first()
 
     assert DUT.get_attributes() == ATTRIBUTES
 
@@ -51,13 +56,13 @@ def test_set_attributes(test_common_dao):
     """ set_attributes() should return a zero error code on success. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKGroup).first()
+    DUT = _session.query(RAMSTKStatus).first()
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 0
-    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKGroup {0:d} "
-                    "attributes.".format(DUT.group_id))
+    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKStatus {0:d} "
+                    "attributes.".format(DUT.status_id))
 
 
 @pytest.mark.integration
@@ -65,14 +70,14 @@ def test_set_attributes_missing_key(test_common_dao):
     """ set_attributes() should return a 40 error code when passed too few attributes. """
     _session = test_common_dao.RAMSTK_SESSION(
         bind=test_common_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKGroup).first()
+    DUT = _session.query(RAMSTKStatus).first()
 
-    ATTRIBUTES.pop('group_type')
+    ATTRIBUTES.pop('name')
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 40
-    assert _msg == ("RAMSTK ERROR: Missing attribute 'group_type' in attribute "
-                    "dictionary passed to RAMSTKGroup.set_attributes().")
+    assert _msg == ("RAMSTK ERROR: Missing attribute 'name' in attribute "
+                    "dictionary passed to RAMSTKStatus.set_attributes().")
 
-    ATTRIBUTES['group_type'] = 'workgroup'
+    ATTRIBUTES['name'] = ''
